// packages/core-sdk/src/abi/generated/factories/RelationshipModule__factory.ts
const _abi = [
  {
    inputs: [
      {
        components: [
          {
            internalType: "address",
            name: "sourceContract",
            type: "address",
          },
          {
            internalType: "uint256",
            name: "sourceId",
            type: "uint256",
          },
          {
            internalType: "address",
            name: "destContract",
            type: "address",
          },
          {
            internalType: "uint256",
            name: "destId",
            type: "uint256",
          },
          {
            internalType: "bytes32",
            name: "relationshipId",
            type: "bytes32",
          },
          {
            internalType: "uint256",
            name: "ttl",
            type: "uint256",
          },
        ],
        internalType: "struct IRelationshipModule.RelationshipParams",
        name: "params",
        type: "tuple",
      },
      {
        internalType: "bytes",
        name: "data",
        type: "bytes",
      },
    ],
    name: "relate",
    outputs: [],
    stateMutability: "nonpayable",
    type: "function",
  },
  {
    inputs: [
      {
        components: [
          {
            internalType: "address",
            name: "sourceContract",
            type: "address",
          },
          {
            internalType: "uint256",
            name: "sourceId",
            type: "uint256",
          },
          {
            internalType: "address",
            name: "destContract",
            type: "address",
          },
          {
            internalType: "uint256",
            name: "destId",
            type: "uint256",
          },
          {
            internalType: "bytes32",
            name: "relationshipId",
            type: "bytes32",
          },
          {
            internalType: "uint256",
            name: "ttl",
            type: "uint256",
          },
        ],
        internalType: "struct IRelationshipModule.RelationshipParams",
        name: "params",
        type: "tuple",
      },
    ],
    name: "unrelate",
    outputs: [],
    stateMutability: "nonpayable",
    type: "function",
  },
  {
    inputs: [
      {
        components: [
          {
            internalType: "address",
            name: "sourceContract",
            type: "address",
          },
          {
            internalType: "uint256",
            name: "sourceId",
            type: "uint256",
          },
          {
            internalType: "address",
            name: "destContract",
            type: "address",
          },
          {
            internalType: "uint256",
            name: "destId",
            type: "uint256",
          },
          {
            internalType: "bytes32",
            name: "relationshipId",
            type: "bytes32",
          },
          {
            internalType: "uint256",
            name: "ttl",
            type: "uint256",
          },
        ],
        internalType: "struct IRelationshipModule.RelationshipParams",
        name: "params",
        type: "tuple",
      },
    ],
    name: "isRelationshipExpired",
    outputs: [
      {
        internalType: "bool",
        name: "",
        type: "bool",
      },
    ],
    stateMutability: "view",
    type: "function",
  },
  {
    inputs: [
      {
        components: [
          {
            internalType: "address",
            name: "sourceContract",
            type: "address",
          },
          {
            internalType: "uint256",
            name: "sourceId",
            type: "uint256",
          },
          {
            internalType: "address",
            name: "destContract",
            type: "address",
          },
          {
            internalType: "uint256",
            name: "destId",
            type: "uint256",
          },
          {
            internalType: "bytes32",
            name: "relationshipId",
            type: "bytes32",
          },
          {
            internalType: "uint256",
            name: "ttl",
            type: "uint256",
          },
        ],
        internalType: "struct IRelationshipModule.RelationshipParams",
        name: "params",
        type: "tuple",
      },
    ],
    name: "areTheyRelated",
    outputs: [
      {
        internalType: "bool",
        name: "",
        type: "bool",
      },
    ],
    stateMutability: "view",
    type: "function",
  },
  {
    inputs: [
      {
        internalType: "string",
        name: "name",
        type: "string",
      },
    ],
    name: "getRelationshipId",
    outputs: [
      {
        internalType: "bytes32",
        name: "",
        type: "bytes32",
      },
    ],
    stateMutability: "pure",
    type: "function",
  },
] as const;

export class RelationshipModule__factory {
  static readonly abi = _abi;
  static createInterface(): RelationshipModuleInterface {
    return new utils.Interface(_abi) as RelationshipModuleInterface;
  }
  static connect(
    address: string,
    signerOrProvider: Signer | Provider
  ): RelationshipModule {
    return new Contract(address, _abi, signerOrProvider) as RelationshipModule;
  }
}